requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pgvector>=0.3.0

# LLM
langchain>=0.3.0
langchain-openai>=0.3.0
openai>=1.50.0

# Media processing
yt-dlp>=2024.0.0

# Utils
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Auth
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.7

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
